requirements.txt
PyPDF2
scikit-learn
google-generativeai
streamlit
